refactor(Gateway.Middlewares): rename namespace for middlewares
- Excluded Middlewares for code coverage
